requirements.txt
python-telegram-bot[job-queue]==21.6
tinydb==4.8.0
